+++ app/experiences/[id]/loading.tsx
@@ -1,0 +1,5 @@
+import { ExperienceDetailSkeleton } from '@/app/components/skeletons/ExperienceDetailSkeleton';
+
+export default function Loading() {
+  return <ExperienceDetailSkeleton />;
+}